Removing deprecated lines from the make force file
Some lines were used during the development stage of this feature and we forgot to remove them when merging to the dev branch. I'm just removing those lines, but also adjusting the destiny grid file to use the ```config_pyroms.txt``` and find the right grid when activating the xesmf interpolation.